+++ .github/release-notes.md
@@ -1,3 +1,4 @@
 * Attempt to fetch missing summaries when viewing player history
 * Added green highlight to players in your game when viewing history
 * Fixed players not being clickable in past game popup
+* Added display for known friends when clicking on player in game